requirements.txt
requests==2.31.0
python-dotenv==1.0.0
Flask==3.0.0
Flask-WTF==1.2.1
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
Werkzeug==3.0.1
Click==8.1.7
Flask-WTF==1.2.1
email-validator==2.1.0.post1
gunicorn==21.2.0
